mvoe tunicwilds/metadata GET to tunicwilds GET
